zcash_pool_migration_memory: extract the in-memory test backends into a crate
Move the in-memory mock backends (MockBackend, CommitMock, LayeredMock) and the
witness/key test helpers (spending_key, regtest_network, single_note_witness,
shared_anchor_witnesses, prep_fee) out of zcash_pool_migration_backend's
zcash_pool_migration_memory, so other crates can reuse them. The crate's docs
state prominently that it is for testing only and not for production.

Because a dev-dependency cycle rebuilds the backend library under cfg(test) as a
distinct crate instance (so mock trait impls built against the plain library do
not satisfy the test binary's same-named traits), the mock-consuming engine
tests move from unit tests into integration tests under
zcash_pool_migration_backend/tests/, which link the same library instance as the
memory crate. This mirrors how zcash_client_memory is consumed. The build-module
tests are unaffected: their helpers cross only shared orchard/pczt types, so they
re-export the moved helpers from the memory crate and keep compiling as unit
tests. The relocated commits_multi_layer test hand-builds a MigrationPlan, so a
public MigrationPlan::from_parts constructor is added (matching the existing
PreparationPlan::from_parts and NoteSplitPlan::from_stored_parts). No production
behavior changes; the test count is unchanged (94).
