+++ zcash_pool_migration_backend/Cargo.toml
@@ -48,6 +48,11 @@ rand_chacha.workspace = true
 # The `local-consensus` feature exposes `LocalNetwork`, used to build the migration transactions
 # against regtest params with the relevant network upgrades (including NU6.3) active.
 zcash_protocol = { workspace = true, features = ["local-consensus"] }
+# Shared, test-only in-memory backends (`MockBackend`, `CommitMock`, `LayeredMock`) and witness
+# helpers the engine tests use. The dev-dependency cycle back onto this crate is fine: dev-deps do
+# not enter the normal build graph (the same relationship zcash_client_memory has with
+# zcash_client_backend).
+zcash_pool_migration_memory = { path = "../zcash_pool_migration_memory" }
 
 [features]
 ## Builds the pure PCZT transaction builder (the `build` module), which turns a note-split plan and
