+++ zcash_pool_migration_backend/src/engine.rs
@@ -1126,773 +1126,3 @@ where
         .map_err(CommitError::Backend)?;
     Ok((state, unsigned))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rand_chacha::ChaCha8Rng;
-    use rand_core::SeedableRng;
-    use zcash_protocol::value::COIN;
-
-    use zcash_protocol::local_consensus::LocalNetwork;
-
-    use crate::preparation::FUNDING_OUTPUTS_PER_TX;
-
-    /// A local network with NU6.3 active at a low height, matching the build test network, so the
-    /// canonical fees and activation checks compute in planning tests.
-    fn test_net() -> LocalNetwork {
-        LocalNetwork {
-            overwinter: Some(BlockHeight::from_u32(1)),
-            sapling: Some(BlockHeight::from_u32(2)),
-            blossom: Some(BlockHeight::from_u32(3)),
-            heartwood: Some(BlockHeight::from_u32(4)),
-            canopy: Some(BlockHeight::from_u32(5)),
-            nu5: Some(BlockHeight::from_u32(6)),
-            nu6: Some(BlockHeight::from_u32(7)),
-            nu6_1: Some(BlockHeight::from_u32(8)),
-            nu6_2: Some(BlockHeight::from_u32(9)),
-            nu6_3: Some(BlockHeight::from_u32(10)),
-            #[cfg(zcash_unstable = "nu7")]
-            nu7: None,
-        }
-    }
-
-    /// The canonical fees on the test network, computed exactly as `plan_migration` computes them.
-    fn test_fees() -> (Zatoshis, Zatoshis) {
-        canonical_fees(&test_net(), BlockHeight::from_u32(2_000_000))
-            .expect("the canonical fees compute")
-    }
-
-    /// A count-only preparation-layout stub for tests that exercise the split in isolation: one
-    /// padded transaction per [`FUNDING_OUTPUTS_PER_TX`] funding notes.
-    fn prep_tx_count_stub(notes: &[Zatoshis]) -> Option<usize> {
-        Some(notes.len().div_ceil(FUNDING_OUTPUTS_PER_TX))
-    }
-
-    /// A minimal in-memory backend: a fixed set of note values and a chain tip.
-    struct MockBackend {
-        notes: Vec<Zatoshis>,
-        tip: BlockHeight,
-        stored: Option<MigrationState>,
-    }
-
-    impl MockBackend {
-        fn new(notes: Vec<u64>, tip: u32) -> Self {
-            MockBackend {
-                notes: notes
-                    .into_iter()
-                    .map(|v| Zatoshis::from_u64(v).expect("test note values are valid"))
-                    .collect(),
-                tip: BlockHeight::from_u32(tip),
-                stored: None,
-            }
-        }
-    }
-
-    impl MigrationBackend for MockBackend {
-        type Error = core::convert::Infallible;
-
-        fn spendable_orchard_note_values(&self) -> Result<Vec<Zatoshis>, Self::Error> {
-            Ok(self.notes.clone())
-        }
-
-        fn chain_tip_height(&self) -> Result<BlockHeight, Self::Error> {
-            Ok(self.tip)
-        }
-    }
-
-    impl PoolMigrationRead for MockBackend {
-        type Error = core::convert::Infallible;
-
-        fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
-            Ok(self.stored.clone())
-        }
-    }
-
-    impl PoolMigrationWrite for MockBackend {
-        fn put_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
-            self.stored = Some(state.clone());
-            Ok(())
-        }
-
-        fn update_transaction(
-            &mut self,
-            id: MigrationTxId,
-            state: MigrationTxState,
-        ) -> Result<(), Self::Error> {
-            if let Some(stored) = &mut self.stored {
-                if let Some(tx) = stored.transactions.iter_mut().find(|t| t.id == id) {
-                    tx.state = state;
-                }
-            }
-            Ok(())
-        }
-    }
-
-    #[test]
-    fn plans_a_migration_from_a_balance() {
-        let backend = MockBackend::new(vec![100 * COIN, 40 * COIN], 2_000_000);
-        let mut rng = ChaCha8Rng::seed_from_u64(1);
-        let plan =
-            plan_migration(&test_net(), &backend, &mut rng).expect("a fundable balance plans");
-
-        // Something is migrated; the schedule has one entry per funding note; the preparation mints
-        // exactly the (reconciled) funding notes; and reconciliation only ever drops, never adds.
-        assert!(!plan.funding_notes().is_empty());
-        assert_eq!(plan.schedule().len(), plan.funding_notes().len());
-
-        // The preparation schedule mirrors the layers' shape, is non-decreasing within each
-        // layer, and each layer starts past the previous one (temporal decoupling with layer
-        // serialization).
-        assert_eq!(plan.prep_schedule().len(), plan.preparation().layer_count());
-        let mut prev_layer_end = BlockHeight::from_u32(2_000_000);
-        for (layer, heights) in plan.preparation().layers().iter().zip(plan.prep_schedule()) {
-            assert_eq!(heights.len(), layer.len());
-            let mut prev = prev_layer_end;
-            for &h in heights {
-                assert!(h >= prev, "prep schedule is non-decreasing across layers");
-                prev = h;
-            }
-            prev_layer_end = prev;
-        }
-
-        // The schedule floor: no transfer is scheduled before the earliest height at which a
-        // candidate anchor boundary exists, given the drawn preparation schedule plus the mining
-        // margin.
-        let est_last_prep = plan
-            .prep_schedule()
-            .last()
-            .and_then(|layer| layer.last())
-            .copied()
-            .unwrap_or(BlockHeight::from_u32(2_000_000))
-            + EST_PREP_LAYER_MINING_BLOCKS;
-        let earliest =
-            crate::scheduling::earliest_broadcast_height(BlockHeight::from_u32(10), est_last_prep);
-        assert!(
-            plan.schedule()
-                .iter()
-                .all(|s| s.broadcast_height() >= earliest),
-            "every scheduled transfer is at or after the anchor-viability floor"
-        );
-        assert_eq!(
-            plan.preparation().funding_notes().len(),
-            plan.funding_notes().len()
-        );
-        assert!(plan.funding_notes().len() <= plan.note_split().migration_outputs().len());
-    }
-
-    /// SHUFFLE (ZIP 318): the crossings are non-increasing, so an unshuffled schedule would
-    /// pair them with non-decreasing broadcast heights and the on-chain temporal sequence of
-    /// denominations would spell out the balance. The drawn permutation makes the heights
-    /// non-monotone in crossing index (deterministic for this seed).
-    #[test]
-    fn transfer_schedule_is_shuffled() {
-        // A balance that splits into several denominations, so the order is observable.
-        let backend = MockBackend::new(vec![78 * COIN], 2_000_000);
-        let mut rng = ChaCha8Rng::seed_from_u64(3);
-        let plan =
-            plan_migration(&test_net(), &backend, &mut rng).expect("a fundable balance plans");
-        assert!(
-            plan.funding_notes().len() >= 4,
-            "the balance splits into several transfers: {}",
-            plan.funding_notes().len()
-        );
-        let heights: Vec<u32> = plan
-            .schedule()
-            .iter()
-            .map(|s| u32::from(s.broadcast_height()))
-            .collect();
-        assert!(
-            heights.windows(2).any(|w| w[0] > w[1]),
-            "the transfer broadcast order is shuffled: {heights:?}"
-        );
-    }
-
-    #[test]
-    fn empty_balance_has_nothing_to_migrate() {
-        let backend = MockBackend::new(Vec::new(), 2_000_000);
-        let mut rng = ChaCha8Rng::seed_from_u64(1);
-        assert!(matches!(
-            plan_migration(&test_net(), &backend, &mut rng),
-            Err(MigrationError::NothingToMigrate)
-        ));
-    }
-
-    #[test]
-    fn stores_loads_and_updates_a_migration() {
-        let mut backend = MockBackend::new(Vec::new(), 0);
-        assert!(backend.get_migration().unwrap().is_none());
-
-        let mut rng = ChaCha8Rng::seed_from_u64(1);
-        let (prep_tx_fee, transfer_buffer) = test_fees();
-        let note_split = crate::note_splitting::plan_note_split(
-            Zatoshis::from_u64(100 * COIN).expect("test balance is valid"),
-            transfer_buffer,
-            prep_tx_fee,
-            &prep_tx_count_stub,
-            &mut rng,
-        );
-        let tx = MigrationTransaction {
-            id: MigrationTxId(0),
-            kind: MigrationTxKind::Transfer { crossing: 0 },
-            pczt: vec![1, 2, 3], // a stand-in for the serialized pre-signed PCZT
-            depends_on: Vec::new(),
-            scheduled_height: BlockHeight::from_u32(2_000_100),
-            expiry_height: BlockHeight::from_u32(2_069_220),
-            anchor_boundary: None,
-            state: MigrationTxState::Signed,
-        };
-        let state = MigrationState {
-            status: MigrationStatus::Committed,
-            note_split,
-            funding_notes: Vec::new(),
-            preparation: crate::preparation::PreparationPlan::from_parts(Vec::new(), Vec::new()),
-            transactions: vec![tx],
-        };
-        backend.put_migration(&state).unwrap();
-
-        // The stored transactions round-trip, and a state update persists.
-        let loaded = backend
-            .get_migration()
-            .unwrap()
-            .expect("a migration is stored");
-        assert_eq!(loaded.status, MigrationStatus::Committed);
-        assert_eq!(loaded.transactions, state.transactions);
-
-        backend
-            .update_transaction(
-                MigrationTxId(0),
-                MigrationTxState::Mined {
-                    height: BlockHeight::from_u32(2_000_105),
-                },
-            )
-            .unwrap();
-        let loaded = backend.get_migration().unwrap().unwrap();
-        assert_eq!(
-            loaded.transactions[0].state,
-            MigrationTxState::Mined {
-                height: BlockHeight::from_u32(2_000_105)
-            }
-        );
-    }
-}
-
-#[cfg(all(test, feature = "orchard"))]
-mod commit_tests {
-    use super::*;
-    use rand_chacha::ChaCha8Rng;
-    use rand_core::SeedableRng;
-    use zcash_protocol::value::COIN;
-
-    use orchard::keys::{FullViewingKey, SpendAuthorizingKey};
-
-    use crate::build::sign_pczt;
-    use crate::build::test_util::{
-        TARGET_HEIGHT, regtest_network, single_note_witness, spending_key,
-    };
-    use crate::note_splitting::NoteSplitPlan;
-    use crate::preparation::{PREP_TX_ACTIONS, plan_preparation};
-
-    /// The canonical fees on the regtest network at the build height, computed exactly as
-    /// `plan_migration` computes them.
-    fn commit_test_fees() -> (Zatoshis, Zatoshis) {
-        canonical_fees(&regtest_network(true), BlockHeight::from_u32(TARGET_HEIGHT))
-            .expect("the canonical fees compute")
-    }
-
-    fn prep_fee() -> Zatoshis {
-        commit_test_fees().0
-    }
-
-    /// A wallet mock holding the account's key and its spendable notes' PLAINTEXTS — nothing
-    /// more: with anchors and witnesses deferred to proving time (ZIP 374), building and signing
-    /// an entire migration needs no tree access at all. It signs with its own spend-authorizing
-    /// key and stores the migration in memory.
-    struct CommitMock {
-        wallet_notes: Vec<orchard::note::Note>,
-        fvk: FullViewingKey,
-        ask: SpendAuthorizingKey,
-        stored: Option<MigrationState>,
-    }
-
-    impl CommitMock {
-        /// A mock wallet holding single notes of the given values, derived from `seed`.
-        fn new(seed: u64, values: &[u64]) -> Self {
-            let sk = spending_key(seed);
-            let fvk = FullViewingKey::from(&sk);
-            let wallet_notes = values
-                .iter()
-                .enumerate()
-                .map(|(i, &v)| single_note_witness(&fvk, v, seed.wrapping_add(i as u64)).0)
-                .collect();
-            CommitMock {
-                wallet_notes,
-                fvk,
-                ask: SpendAuthorizingKey::from(&sk),
-                stored: None,
-            }
-        }
-    }
-
-    impl MigrationBackend for CommitMock {
-        type Error = core::convert::Infallible;
-
-        fn spendable_orchard_note_values(&self) -> Result<Vec<Zatoshis>, Self::Error> {
-            Ok(self
-                .wallet_notes
-                .iter()
-                .map(|n| Zatoshis::from_u64(n.value().inner()).expect("test note values are valid"))
-                .collect())
-        }
-
-        fn chain_tip_height(&self) -> Result<BlockHeight, Self::Error> {
-            Ok(BlockHeight::from_u32(2_000_000))
-        }
-    }
-
-    impl PoolMigrationRead for CommitMock {
-        type Error = core::convert::Infallible;
-
-        fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
-            Ok(self.stored.clone())
-        }
-    }
-
-    impl PoolMigrationWrite for CommitMock {
-        fn put_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
-            self.stored = Some(state.clone());
-            Ok(())
-        }
-
-        fn update_transaction(
-            &mut self,
-            id: MigrationTxId,
-            state: MigrationTxState,
-        ) -> Result<(), Self::Error> {
-            if let Some(stored) = &mut self.stored {
-                if let Some(tx) = stored.transactions.iter_mut().find(|t| t.id == id) {
-                    tx.state = state;
-                }
-            }
-            Ok(())
-        }
-    }
-
-    impl MigrationCrypto for CommitMock {
-        type Error = core::convert::Infallible;
-
-        fn orchard_fvk(&self) -> Result<FullViewingKey, Self::Error> {
-            Ok(self.fvk.clone())
-        }
-
-        fn resolve_wallet_note(&self, index: usize) -> Result<orchard::note::Note, Self::Error> {
-            Ok(self.wallet_notes[index])
-        }
-
-        fn sign(&self, pczt: pczt::Pczt) -> Result<pczt::Pczt, Self::Error> {
-            Ok(sign_pczt(pczt, &self.ask).expect("signs the migration PCZT"))
-        }
-    }
-
-    /// A planned single-note migration and the mock wallet that holds the note.
-    fn single_note_setup(seed: u64, balance: u64) -> (CommitMock, MigrationPlan) {
-        let backend = CommitMock::new(seed, &[balance]);
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        let plan = plan_migration(&regtest_network(true), &backend, &mut rng)
-            .expect("a fundable balance plans");
-        (backend, plan)
-    }
-
-    /// The WHOLE migration — every preparation transaction and every transfer — is built and
-    /// SIGNED in the one commit pass, before anything is broadcast or mined: the funding notes
-    /// are recovered from the built preparation bundles, and every stored PCZT carries ABSENT
-    /// anchors (ZIP 374), to be installed at proving time against each transaction's anchor.
-    #[test]
-    fn commits_the_whole_migration_in_one_pass() {
-        let seed = 7u64;
-        let (mut backend, plan) = single_note_setup(seed, 78 * COIN);
-        // A single note funding a handful of denominations needs one preparation layer.
-        assert_eq!(plan.preparation().layers().len(), 1);
-        let params = regtest_network(true);
-        let prep_count: usize = plan.preparation().layers().iter().map(|l| l.len()).sum();
-        let transfer_count = plan.funding_notes().len();
-        assert!(transfer_count >= 2, "several transfers: {transfer_count}");
-
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 1);
-        let state = commit_preparation(
-            &params,
-            BlockHeight::from_u32(TARGET_HEIGHT),
-            &mut backend,
-            &plan,
-            &mut rng,
-        )
-        .expect("commits the migration");
-        assert_eq!(state.status, MigrationStatus::Committed);
-        assert_eq!(state.transactions.len(), prep_count + transfer_count);
-
-        // The funding notes exist only once the last preparation transaction has mined; every
-        // drawn boundary must lie at or after that estimate, exactly as the draw is floored.
-        let est_last_prep = state
-            .transactions
-            .iter()
-            .filter(|t| matches!(t.kind, MigrationTxKind::Preparation { .. }))
-            .map(|t| t.scheduled_height)
-            .max()
-            .expect("the plan has a preparation")
-            + EST_PREP_LAYER_MINING_BLOCKS;
-
-        for tx in &state.transactions {
-            // ONE signing phase: everything is signed at commit, before anything mines.
-            assert_eq!(tx.state, MigrationTxState::Signed, "signed at commit");
-            assert!(!tx.pczt.is_empty());
-            let parsed = pczt::Pczt::parse(&tx.pczt).expect("the stored PCZT parses");
-            // Anchors are deferred (ZIP 374): every stored PCZT carries ABSENT anchors, and the
-            // pre-signature commits to the stored canonical expiry for the drawn schedule.
-            assert!(parsed.orchard().anchor().is_none());
-            assert!(parsed.ironwood().anchor().is_none());
-            assert_eq!(
-                *parsed.global().expiry_height(),
-                u32::from(tx.expiry_height),
-                "the embedded expiry matches the stored schedule expiry"
-            );
-            match tx.kind {
-                MigrationTxKind::Preparation { .. } => {
-                    assert!(
-                        tx.depends_on.is_empty(),
-                        "single-layer preps are independent"
-                    );
-                    assert!(tx.anchor_boundary.is_none());
-                }
-                MigrationTxKind::Transfer { .. } => {
-                    assert!(
-                        !tx.depends_on.is_empty(),
-                        "a transfer BROADCASTS only after the preparation mines"
-                    );
-                    // The boundary anchor the transfer will PROVE against is drawn at commit
-                    // time: on the boundary grid, strictly above the NU6.3 activation, at or
-                    // after the estimated height the last preparation has mined, and strictly
-                    // below the most recent boundary at the scheduled broadcast height.
-                    let b = u32::from(
-                        tx.anchor_boundary
-                            .expect("every transfer carries its boundary"),
-                    );
-                    assert_eq!(b % crate::scheduling::BOUNDARY_MODULUS, 0);
-                    assert!(b > 10, "boundary above the regtest NU6.3 activation");
-                    assert!(
-                        b >= u32::from(est_last_prep).div_ceil(crate::scheduling::BOUNDARY_MODULUS)
-                            * crate::scheduling::BOUNDARY_MODULUS
-                    );
-                    assert!(
-                        b < u32::from(crate::scheduling::most_recent_boundary(tx.scheduled_height))
-                    );
-                }
-            }
-        }
-        assert!(backend.get_migration().unwrap().is_some());
-    }
-
-    /// A lone whale fanning out into more funding notes than one transaction holds needs a
-    /// MULTI-LAYER preparation — and it still signs in the SAME single pass: the later layer's
-    /// feeder spends and the transfers' funding notes are recovered from the earlier layers'
-    /// built (unmined) bundles. Mining then gates only the broadcast order, which the state
-    /// machine walks layer by layer.
-    #[test]
-    fn commits_a_multi_layer_migration_in_one_pass() {
-        let seed = 11u64;
-        let sk = spending_key(seed);
-        let fvk = FullViewingKey::from(&sk);
-
-        // 15 funding notes (one more than a single transaction's FUNDING_OUTPUTS_PER_TX) force a
-        // two-layer balanced fan-out. Each is a valid self-funding note (a crossing value plus
-        // the transfer fee buffer), so its transfer balances.
-        let buffer = u64::from(commit_test_fees().1);
-        let crossing = COIN; // 1 ZEC crossing per note
-        let funding_note = crossing + buffer;
-        let funding: Vec<u64> = core::iter::repeat_n(funding_note, 15).collect();
-
-        // A whale generously larger than the balanced-tree cost, so the fan-out fast path
-        // triggers.
-        let whale = funding.iter().sum::<u64>() + 16 * u64::from(prep_fee());
-        let whale_zats = [Zatoshis::from_u64(whale).expect("test whale is valid")];
-        let funding_zats: Vec<Zatoshis> = funding
-            .iter()
-            .map(|&v| Zatoshis::from_u64(v).expect("test funding values are valid"))
-            .collect();
-        let preparation = plan_preparation(&whale_zats, &funding_zats, prep_fee())
-            .expect("a fundable whale plans");
-        assert_eq!(
-            preparation.layers().len(),
-            2,
-            "15 funding notes fan out across two layers"
-        );
-
-        // A note split whose outputs are the funding notes and whose crossings are those less
-        // the buffer, so each transfer crosses one ZEC.
-        let crossings: Vec<u64> = funding.iter().map(|&f| f - buffer).collect();
-        let note_split = NoteSplitPlan::from_stored_parts(
-            crossings
-                .iter()
-                .map(|&v| Zatoshis::from_u64(v).expect("test crossings are valid"))
-                .collect(),
-            Zatoshis::from_u64(buffer).expect("the buffer is valid"),
-            None,
-            Zatoshis::from_u64(preparation.transaction_count() as u64 * u64::from(prep_fee()))
-                .expect("the reserved fees are valid"),
-            whale_zats[0],
-            Zatoshis::from_u64(crossings.iter().sum()).expect("the crossing total is valid"),
-        )
-        .expect("a consistent stored plan reconstructs");
-        let mut rng = ChaCha8Rng::seed_from_u64(seed);
-        // A drawn preparation schedule in the layers' shape, each layer based past the previous
-        // one, exactly as `plan_migration` draws it.
-        let mut prep_schedule: Vec<Vec<BlockHeight>> = Vec::new();
-        let mut layer_base = BlockHeight::from_u32(2_000_000);
-        for layer in preparation.layers() {
-            let heights = crate::scheduling::schedule_prep_broadcast_heights(
-                layer_base,
-                layer.len(),
-                &mut rng,
-            );
-            layer_base =
-                heights.last().copied().unwrap_or(layer_base) + EST_PREP_LAYER_MINING_BLOCKS;
-            prep_schedule.push(heights);
-        }
-        // Floor the transfer schedule so every transfer has a candidate anchor boundary above
-        // the estimated last-preparation mining height, exactly as `plan_migration` floors it.
-        let nu63_activation = {
-            use zcash_protocol::consensus::{NetworkUpgrade, Parameters as _};
-            regtest_network(true)
-                .activation_height(NetworkUpgrade::Nu6_3)
-                .expect("NU6.3 is active on the test network")
-        };
-        let schedule_base =
-            crate::scheduling::earliest_broadcast_height(nu63_activation, layer_base);
-        let schedule = crate::scheduling::schedule(schedule_base, funding.len(), &mut rng);
-        let plan = MigrationPlan {
-            note_split,
-            funding_notes: funding_zats,
-            preparation,
-            prep_schedule,
-            schedule,
-        };
-
-        let mut backend = CommitMock {
-            wallet_notes: vec![single_note_witness(&fvk, whale, seed).0],
-            fvk,
-            ask: SpendAuthorizingKey::from(&sk),
-            stored: None,
-        };
-        let params = regtest_network(true);
-        let prep_count = plan.preparation().transaction_count();
-        let transfer_count = funding.len();
-
-        // ONE pass builds and signs both layers and every transfer.
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 1);
-        let state = commit_preparation(
-            &params,
-            BlockHeight::from_u32(TARGET_HEIGHT),
-            &mut backend,
-            &plan,
-            &mut rng,
-        )
-        .expect("commits the migration");
-        assert_eq!(state.transactions.len(), prep_count + transfer_count);
-        for tx in &state.transactions {
-            assert_eq!(tx.state, MigrationTxState::Signed, "signed at commit");
-            assert!(!tx.pczt.is_empty());
-        }
-        let layer0_ids: Vec<MigrationTxId> = state
-            .transactions
-            .iter()
-            .filter(|t| matches!(t.kind, MigrationTxKind::Preparation { layer: 0, .. }))
-            .map(|t| t.id)
-            .collect();
-        assert_eq!(layer0_ids.len(), 1, "one root transaction in layer 0");
-        for tx in &state.transactions {
-            match tx.kind {
-                MigrationTxKind::Preparation { layer, .. } if layer > 0 => {
-                    assert_eq!(
-                        tx.depends_on, layer0_ids,
-                        "a later layer broadcasts only after its predecessor mines"
-                    );
-                }
-                _ => {}
-            }
-        }
-
-        // The state machine walks the broadcasts in dependency order: layer 0 first; layer 1
-        // only once layer 0 mines; the transfers only once the whole preparation mines.
-        let mut state = state;
-        let target = BlockHeight::from_u32(2_100_000);
-        match state.next_step(target) {
-            crate::state::AdvanceStep::Broadcast { id } => {
-                assert!(layer0_ids.contains(&id), "layer 0 broadcasts first")
-            }
-            other => panic!("expected a broadcast step, got {other:?}"),
-        }
-        for id in &layer0_ids {
-            state.mark_mined(*id, BlockHeight::from_u32(2_000_010));
-        }
-        let layer1_ids: Vec<MigrationTxId> = state
-            .transactions
-            .iter()
-            .filter(|t| matches!(t.kind, MigrationTxKind::Preparation { layer: 1, .. }))
-            .map(|t| t.id)
-            .collect();
-        match state.next_step(target) {
-            crate::state::AdvanceStep::Broadcast { id } => {
-                assert!(
-                    layer1_ids.contains(&id),
-                    "layer 1 broadcasts once layer 0 mines"
-                )
-            }
-            other => panic!("expected a broadcast step, got {other:?}"),
-        }
-        for id in &layer1_ids {
-            state.mark_mined(*id, BlockHeight::from_u32(2_000_020));
-        }
-        match state.next_step(target) {
-            crate::state::AdvanceStep::Broadcast { id } => {
-                let tx = state
-                    .transactions
-                    .iter()
-                    .find(|t| t.id == id)
-                    .expect("the step names a stored transaction");
-                assert!(
-                    matches!(tx.kind, MigrationTxKind::Transfer { .. }),
-                    "the transfers broadcast once the whole preparation mines"
-                );
-            }
-            other => panic!("expected a broadcast step, got {other:?}"),
-        }
-    }
-
-    /// The EXTERNAL path builds the whole migration unsigned in the same one pass, and the
-    /// unsigned transactions split into signing sessions bounded by the device's action budget —
-    /// consecutive topological prefixes, never gated on mining.
-    #[test]
-    fn external_signing_batches_by_action_budget() {
-        let seed = 19u64;
-        let (mut backend, plan) = single_note_setup(seed, 78 * COIN);
-        let params = regtest_network(true);
-
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 1);
-        let (mut state, unsigned) = build_preparation_unsigned(
-            &params,
-            BlockHeight::from_u32(TARGET_HEIGHT),
-            &mut backend,
-            &plan,
-            &mut rng,
-        )
-        .expect("builds the migration unsigned");
-        assert_eq!(unsigned.len(), state.transactions.len());
-        for tx in &state.transactions {
-            assert_eq!(tx.state, MigrationTxState::AwaitingSignature);
-        }
-
-        // Sessions are consecutive prefixes bounded by the action budget; a preparation is
-        // PREP_TX_ACTIONS actions, so a budget of one preparation plus one transfer splits the
-        // list without ever exceeding the budget (every batch is non-empty and within budget).
-        let budget = PREP_TX_ACTIONS
-            + crate::note_splitting::SOURCE_ACTIONS_PER_TRANSFER
-            + crate::note_splitting::DESTINATION_ACTIONS_PER_TRANSFER;
-        let total = unsigned.len();
-        let sessions = batch_unsigned_by_action_budget(unsigned, budget);
-        assert!(sessions.len() > 1, "several sessions: {}", sessions.len());
-        assert_eq!(sessions.iter().map(|s| s.len()).sum::<usize>(), total);
-        for session in &sessions {
-            assert!(!session.is_empty());
-            assert!(session.iter().map(|tx| tx.actions()).sum::<usize>() <= budget);
-        }
-
-        // Sign every session out of band and apply the signatures back; the whole migration is
-        // then Signed without anything having been broadcast or mined.
-        let ask = SpendAuthorizingKey::from(&spending_key(seed));
-        for session in sessions {
-            for unsigned_tx in session {
-                let (id, bytes) = unsigned_tx.into_parts();
-                let signed = sign_pczt(
-                    pczt::Pczt::parse(&bytes).expect("the unsigned PCZT parses"),
-                    &ask,
-                )
-                .expect("the device signs the transaction");
-                assert!(state.apply_signature(id, signed.serialize().expect("serializes")));
-            }
-        }
-        backend.put_migration(&state).unwrap();
-        for tx in &state.transactions {
-            assert_eq!(tx.state, MigrationTxState::Signed);
-        }
-    }
-
-    /// A committed migration must be resumed, never rebuilt over: a second commit while the
-    /// stored migration is non-terminal is refused (its pre-signed transactions may already be
-    /// broadcast, and a rebuilt migration would double-spend the same notes); a terminal
-    /// (failed/cancelled) migration may be replaced.
-    #[test]
-    fn commit_preparation_refuses_to_overwrite_a_live_migration() {
-        let seed = 13u64;
-        let (mut backend, plan) = single_note_setup(seed, 78 * COIN);
-        let params = regtest_network(true);
-
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 1);
-        commit_preparation(
-            &params,
-            BlockHeight::from_u32(TARGET_HEIGHT),
-            &mut backend,
-            &plan,
-            &mut rng,
-        )
-        .expect("commits the migration");
-
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 2);
-        assert!(matches!(
-            commit_preparation(
-                &params,
-                BlockHeight::from_u32(TARGET_HEIGHT),
-                &mut backend,
-                &plan,
-                &mut rng,
-            ),
-            Err(CommitError::MigrationInProgress)
-        ));
-
-        // A terminal (cancelled) migration may be replaced.
-        let mut stored = backend.get_migration().unwrap().expect("stored");
-        stored.status = MigrationStatus::Failed;
-        backend.put_migration(&stored).unwrap();
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 3);
-        commit_preparation(
-            &params,
-            BlockHeight::from_u32(TARGET_HEIGHT),
-            &mut backend,
-            &plan,
-            &mut rng,
-        )
-        .expect("replaces a terminal migration");
-    }
-
-    /// A wallet note is resolved by its index into the CURRENT spendable set; if that set
-    /// changed since planning (the value at the planned index no longer matches), the commit
-    /// reports a stale plan instead of building against the wrong note.
-    #[test]
-    fn commit_preparation_detects_a_stale_plan() {
-        let seed = 17u64;
-        let (_, plan) = single_note_setup(seed, 78 * COIN);
-
-        // The spendable set shifted between planning and commit: the note at index 0 now has a
-        // different value than the plan recorded.
-        let mut backend = CommitMock::new(seed, &[78 * COIN + COIN]);
-        let mut rng = ChaCha8Rng::seed_from_u64(seed + 1);
-        assert!(matches!(
-            commit_preparation(
-                &regtest_network(true),
-                BlockHeight::from_u32(TARGET_HEIGHT),
-                &mut backend,
-                &plan,
-                &mut rng,
-            ),
-            Err(CommitError::StalePlan)
-        ));
-    }
-}
