+++ zcash_pool_migration_memory/src/lib.rs
@@ -1,0 +1,469 @@
+//! In-memory implementations of the [`zcash_pool_migration_backend`] engine traits, **FOR TESTING
+//! ONLY**.
+//!
+//! This crate provides in-memory implementations of the pool-migration engine traits
+//! ([`MigrationBackend`], [`PoolMigrationRead`], [`PoolMigrationWrite`], and [`MigrationCrypto`]) for
+//! TESTING ONLY; it is not intended for production use. The mock backends hold their notes,
+//! witnesses, and migration state in memory, sign with a hand-held Orchard spend-authorizing key, and
+//! deterministically derive keys and note-commitment-tree witnesses from a seed. None of this is a
+//! substitute for a real wallet backend.
+//!
+//! It mirrors how [`zcash_client_memory`] relates to `zcash_client_backend`: a shared, test-support
+//! crate so several test suites can reuse the same mock implementations.
+//!
+//! [`MigrationBackend`]: zcash_pool_migration_backend::engine::MigrationBackend
+//! [`PoolMigrationRead`]: zcash_pool_migration_backend::engine::PoolMigrationRead
+//! [`PoolMigrationWrite`]: zcash_pool_migration_backend::engine::PoolMigrationWrite
+//! [`MigrationCrypto`]: zcash_pool_migration_backend::engine::MigrationCrypto
+//! [`zcash_client_memory`]: https://docs.rs/zcash_client_memory
+
+use std::cell::RefCell;
+
+use incrementalmerkletree::{Hashable, Level};
+use orchard::keys::{FullViewingKey, Scope, SpendAuthorizingKey, SpendingKey};
+use orchard::note::{ExtractedNoteCommitment, Note, NoteVersion, RandomSeed, Rho};
+use orchard::tree::{MerkleHashOrchard, MerklePath};
+use orchard::value::NoteValue;
+use orchard::{Anchor, NOTE_COMMITMENT_TREE_DEPTH};
+use rand_chacha::ChaCha8Rng;
+use rand_core::{RngCore, SeedableRng};
+use zcash_protocol::consensus::BlockHeight;
+use zcash_protocol::local_consensus::LocalNetwork;
+
+use zcash_pool_migration_backend::build::sign_pczt;
+use zcash_pool_migration_backend::engine::{
+    MigrationBackend, MigrationCrypto, MigrationState, MigrationTxId, MigrationTxState,
+    PoolMigrationRead, PoolMigrationWrite,
+};
+use zcash_pool_migration_backend::note_splitting::{FeePolicy, Zip317FeePolicy};
+use zcash_pool_migration_backend::preparation::PREP_TX_ACTIONS;
+
+/// A post-NU6.3 height (past the regtest NU6.3 activation) at which the migration transactions are
+/// built and their fees computed.
+pub const TARGET_HEIGHT: u32 = 100;
+
+/// The ZIP-317 fee of a padded preparation transaction, as the engine's caller would compute it.
+/// Shared by the migration tests so the mock backends and their scenarios agree on the fee.
+pub fn prep_fee() -> u64 {
+    PREP_TX_ACTIONS as u64 * Zip317FeePolicy.marginal_fee_zatoshi()
+}
+
+/// 32 random bytes from a `seed`-derived RNG, keeping calls deterministic per case.
+fn draw_bytes(rng: &mut ChaCha8Rng) -> [u8; 32] {
+    let mut bytes = [0u8; 32];
+    rng.fill_bytes(&mut bytes);
+    bytes
+}
+
+/// An account's Orchard spending key, derived from `seed` so tests can vary the account across
+/// proptest cases. Draws bytes until they form a valid spending key.
+pub fn spending_key(seed: u64) -> SpendingKey {
+    let mut rng = ChaCha8Rng::seed_from_u64(seed);
+    loop {
+        let bytes = draw_bytes(&mut rng);
+        if let Some(sk) = SpendingKey::from_bytes(bytes).into_option() {
+            return sk;
+        }
+    }
+}
+
+/// A regtest network with the pre-NU6.3 upgrades active, and NU6.3 active only when requested.
+/// The migration builds on a network where NU6.3 (the Ironwood pool) is live.
+pub fn regtest_network(nu6_3_active: bool) -> LocalNetwork {
+    let nu6_3 = if nu6_3_active {
+        Some(BlockHeight::from_u32(10))
+    } else {
+        None
+    };
+    LocalNetwork {
+        overwinter: Some(BlockHeight::from_u32(1)),
+        sapling: Some(BlockHeight::from_u32(2)),
+        blossom: Some(BlockHeight::from_u32(3)),
+        heartwood: Some(BlockHeight::from_u32(4)),
+        canopy: Some(BlockHeight::from_u32(5)),
+        nu5: Some(BlockHeight::from_u32(6)),
+        nu6: Some(BlockHeight::from_u32(7)),
+        nu6_1: Some(BlockHeight::from_u32(8)),
+        nu6_2: Some(BlockHeight::from_u32(9)),
+        nu6_3,
+        #[cfg(zcash_unstable = "nu7")]
+        nu7: None,
+    }
+}
+
+/// An Orchard note of `value` owned by `fvk`, with its randomness drawn from `rng`.
+fn orchard_note(fvk: &FullViewingKey, value: u64, rng: &mut ChaCha8Rng) -> Note {
+    let recipient = fvk.address_at(0u32, Scope::External);
+    let note_value = NoteValue::from_raw(value);
+    let rho = loop {
+        let bytes = draw_bytes(rng);
+        if let Some(rho) = Rho::from_bytes(&bytes).into_option() {
+            break rho;
+        }
+    };
+    let rseed = loop {
+        let bytes = draw_bytes(rng);
+        if let Some(rseed) = RandomSeed::from_bytes(bytes, &rho).into_option() {
+            break rseed;
+        }
+    };
+    Note::from_parts(recipient, note_value, rho, rseed, NoteVersion::V2)
+        .into_option()
+        .expect("valid note parts")
+}
+
+/// An Orchard note of `value` owned by `fvk`, with its randomness derived from `seed` (so
+/// proptest varies the note across cases), placed as the sole leaf of an otherwise-empty
+/// note-commitment tree, with the matching anchor. The authentication path uses the empty-subtree
+/// roots for a single leaf at position 0, so `add_orchard_spend`'s anchor check
+/// (`path.root(cmx) == anchor`) accepts it.
+pub fn single_note_witness(
+    fvk: &FullViewingKey,
+    value: u64,
+    seed: u64,
+) -> (Note, MerklePath, Anchor) {
+    let mut rng = ChaCha8Rng::seed_from_u64(seed);
+    let note = orchard_note(fvk, value, &mut rng);
+    let cmx = ExtractedNoteCommitment::from(note.commitment());
+    let auth_path = core::array::from_fn(|level| {
+        let level = Level::from(level as u8);
+        MerkleHashOrchard::empty_root(level)
+    });
+    let path = MerklePath::from_parts(0, auth_path);
+    let anchor = path.root(cmx);
+    (note, path, anchor)
+}
+
+/// `values.len()` Orchard notes owned by `fvk`, placed as the first leaves of one note-commitment
+/// tree, each paired with its authentication path against the single shared anchor. Randomness is
+/// derived from `seed`. This lets a test build a multi-spend transaction whose spends all anchor to
+/// the same root. Assumes `values` is non-empty.
+pub fn shared_anchor_witnesses(
+    fvk: &FullViewingKey,
+    values: &[u64],
+    seed: u64,
+) -> (Vec<(Note, MerklePath)>, Anchor) {
+    // The notes and their leaf hashes.
+    let notes: Vec<Note> = values
+        .iter()
+        .enumerate()
+        .map(|(i, &v)| {
+            let mut rng = ChaCha8Rng::seed_from_u64(seed.wrapping_add(i as u64));
+            orchard_note(fvk, v, &mut rng)
+        })
+        .collect();
+    let leaves: Vec<MerkleHashOrchard> = notes
+        .iter()
+        .map(|n| MerkleHashOrchard::from_cmx(&ExtractedNoteCommitment::from(n.commitment())))
+        .collect();
+
+    // Filled subtree roots per level: `levels[l][p]` is the root of the subtree at level `l`,
+    // position `p`. Level 0 is the leaves; each higher level combines pairs, using the
+    // empty-subtree root for a missing right sibling. Positions past `levels[l].len()` are empty.
+    let mut levels: Vec<Vec<MerkleHashOrchard>> =
+        Vec::with_capacity(NOTE_COMMITMENT_TREE_DEPTH + 1);
+    levels.push(leaves);
+    for l in 0..NOTE_COMMITMENT_TREE_DEPTH {
+        let level = Level::from(l as u8);
+        let cur = &levels[l];
+        let mut next = Vec::with_capacity(cur.len().div_ceil(2));
+        let mut p = 0;
+        while p < cur.len() {
+            let left = cur[p];
+            let right = cur
+                .get(p + 1)
+                .copied()
+                .unwrap_or_else(|| MerkleHashOrchard::empty_root(level));
+            next.push(MerkleHashOrchard::combine(level, &left, &right));
+            p += 2;
+        }
+        levels.push(next);
+    }
+
+    // Each leaf's authentication path: the sibling subtree root at every level (its computed value
+    // when filled, else the empty-subtree root).
+    let witnesses: Vec<(Note, MerklePath)> = notes
+        .into_iter()
+        .enumerate()
+        .map(|(i, note)| {
+            let auth_path = core::array::from_fn(|l| {
+                let level = Level::from(l as u8);
+                let sibling = (i >> l) ^ 1;
+                levels[l]
+                    .get(sibling)
+                    .copied()
+                    .unwrap_or_else(|| MerkleHashOrchard::empty_root(level))
+            });
+            (note, MerklePath::from_parts(i as u32, auth_path))
+        })
+        .collect();
+
+    // Every leaf's path yields the same root; assert it so a broken helper fails loudly rather
+    // than silently producing mismatched anchors.
+    let root_of = |(note, path): &(Note, MerklePath)| {
+        path.root(ExtractedNoteCommitment::from(note.commitment()))
+    };
+    let anchor = root_of(&witnesses[0]);
+    for w in &witnesses {
+        assert_eq!(root_of(w), anchor, "shared-anchor witnesses inconsistent");
+    }
+    (witnesses, anchor)
+}
+
+/// A minimal in-memory backend: a fixed set of note values and a chain tip. Implements the planning
+/// traits ([`MigrationBackend`], [`PoolMigrationRead`], [`PoolMigrationWrite`]); it holds no keys and
+/// cannot sign, so it is used for the plan/store tests, not the commit tests.
+pub struct MockBackend {
+    notes: Vec<u64>,
+    tip: u32,
+    stored: Option<MigrationState>,
+}
+
+impl MockBackend {
+    /// A backend offering `notes` as the spendable Orchard note values and `tip` as the chain-tip
+    /// height, with no migration stored yet.
+    pub fn new(notes: Vec<u64>, tip: u32) -> Self {
+        MockBackend {
+            notes,
+            tip,
+            stored: None,
+        }
+    }
+}
+
+impl MigrationBackend for MockBackend {
+    type Error = core::convert::Infallible;
+
+    fn spendable_orchard_note_values(&self) -> Result<Vec<u64>, Self::Error> {
+        Ok(self.notes.clone())
+    }
+
+    fn chain_tip_height(&self) -> Result<u32, Self::Error> {
+        Ok(self.tip)
+    }
+}
+
+impl PoolMigrationRead for MockBackend {
+    type Error = core::convert::Infallible;
+
+    fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
+        Ok(self.stored.clone())
+    }
+}
+
+impl PoolMigrationWrite for MockBackend {
+    fn put_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
+        self.stored = Some(state.clone());
+        Ok(())
+    }
+
+    fn update_transaction(
+        &mut self,
+        id: MigrationTxId,
+        state: MigrationTxState,
+    ) -> Result<(), Self::Error> {
+        if let Some(stored) = &mut self.stored {
+            if let Some(tx) = stored.transactions.iter_mut().find(|t| t.id == id) {
+                tx.state = state;
+            }
+        }
+        Ok(())
+    }
+}
+
+/// A wallet holding the account's key and a set of note witnesses against one anchor: index 0 is the
+/// source note the preparation spends, and the rest are the funding notes the transfers spend. It
+/// signs with its own spend-authorizing key and stores the migration in memory. All fields are public
+/// so a test can construct it directly for a specific scenario.
+pub struct CommitMock {
+    /// The spendable Orchard note values reported to the planner.
+    pub notes: Vec<u64>,
+    /// The witnessed notes: index 0 is the preparation's source note, the rest are funding notes.
+    pub witnesses: Vec<(Note, MerklePath)>,
+    /// The shared anchor every witness in `witnesses` was built against.
+    pub anchor: Anchor,
+    /// The account's Orchard full viewing key.
+    pub fvk: FullViewingKey,
+    /// The account's Orchard spend-authorizing key, used to sign the migration PCZTs.
+    pub ask: SpendAuthorizingKey,
+    /// The in-memory migration state (`None` until a migration is committed).
+    pub stored: Option<MigrationState>,
+}
+
+impl MigrationBackend for CommitMock {
+    type Error = core::convert::Infallible;
+
+    fn spendable_orchard_note_values(&self) -> Result<Vec<u64>, Self::Error> {
+        Ok(self.notes.clone())
+    }
+
+    fn chain_tip_height(&self) -> Result<u32, Self::Error> {
+        Ok(2_000_000)
+    }
+}
+
+impl PoolMigrationRead for CommitMock {
+    type Error = core::convert::Infallible;
+
+    fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
+        Ok(self.stored.clone())
+    }
+}
+
+impl PoolMigrationWrite for CommitMock {
+    fn put_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
+        self.stored = Some(state.clone());
+        Ok(())
+    }
+
+    fn update_transaction(
+        &mut self,
+        _id: MigrationTxId,
+        _state: MigrationTxState,
+    ) -> Result<(), Self::Error> {
+        Ok(())
+    }
+}
+
+impl MigrationCrypto for CommitMock {
+    type Error = core::convert::Infallible;
+
+    fn orchard_fvk(&self) -> Result<FullViewingKey, Self::Error> {
+        Ok(self.fvk.clone())
+    }
+
+    fn orchard_anchor(&self) -> Result<Anchor, Self::Error> {
+        Ok(self.anchor)
+    }
+
+    fn resolve_wallet_note(
+        &self,
+        index: usize,
+        _anchor: Anchor,
+    ) -> Result<(Note, MerklePath), Self::Error> {
+        Ok(self.witnesses[index].clone())
+    }
+
+    fn resolve_funding_notes(
+        &self,
+        values: &[u64],
+        _anchor: Anchor,
+    ) -> Result<Vec<(Note, MerklePath)>, Self::Error> {
+        // The funding notes are the witnesses after the source note (index 0).
+        Ok(self.witnesses[1..1 + values.len()].to_vec())
+    }
+
+    fn sign(&self, pczt: pczt::Pczt) -> Result<pczt::Pczt, Self::Error> {
+        Ok(sign_pczt(pczt, &self.ask).expect("signs the migration PCZT"))
+    }
+}
+
+/// A wallet mock for the MULTI-LAYER preparation test. The first `n_wallet` witnesses are the
+/// wallet notes (resolved by index for layer 0); the rest are the notes later layers and the
+/// transfers mint (feeders, then funding notes), resolved by value. A persistent `used` set models
+/// a real wallet where a spent note leaves the spendable set, so a feeder consumed by one layer is
+/// never handed to a later resolution again. The store is real, so mining a layer persists. All
+/// fields are public so a test can construct it directly for a specific scenario.
+pub struct LayeredMock {
+    /// The number of leading witnesses that are wallet notes (resolved by index for layer 0).
+    pub n_wallet: usize,
+    /// The witnessed notes: the wallet notes first, then the minted feeders and funding notes.
+    pub witnesses: Vec<(Note, MerklePath)>,
+    /// A persistent per-witness used-set, so a note consumed by one resolution is never reused.
+    pub used: RefCell<Vec<bool>>,
+    /// The shared anchor every witness in `witnesses` was built against.
+    pub anchor: Anchor,
+    /// The account's Orchard full viewing key.
+    pub fvk: FullViewingKey,
+    /// The account's Orchard spend-authorizing key, used to sign the migration PCZTs.
+    pub ask: SpendAuthorizingKey,
+    /// The in-memory migration state (`None` until a migration is committed).
+    pub stored: Option<MigrationState>,
+}
+
+impl MigrationBackend for LayeredMock {
+    type Error = core::convert::Infallible;
+
+    fn spendable_orchard_note_values(&self) -> Result<Vec<u64>, Self::Error> {
+        Ok(self.witnesses[..self.n_wallet]
+            .iter()
+            .map(|(n, _)| n.value().inner())
+            .collect())
+    }
+
+    fn chain_tip_height(&self) -> Result<u32, Self::Error> {
+        Ok(2_000_000)
+    }
+}
+
+impl PoolMigrationRead for LayeredMock {
+    type Error = core::convert::Infallible;
+
+    fn get_migration(&self) -> Result<Option<MigrationState>, Self::Error> {
+        Ok(self.stored.clone())
+    }
+}
+
+impl PoolMigrationWrite for LayeredMock {
+    fn put_migration(&mut self, state: &MigrationState) -> Result<(), Self::Error> {
+        self.stored = Some(state.clone());
+        Ok(())
+    }
+
+    fn update_transaction(
+        &mut self,
+        id: MigrationTxId,
+        state: MigrationTxState,
+    ) -> Result<(), Self::Error> {
+        if let Some(stored) = &mut self.stored {
+            if let Some(tx) = stored.transactions.iter_mut().find(|t| t.id == id) {
+                tx.state = state;
+            }
+        }
+        Ok(())
+    }
+}
+
+impl MigrationCrypto for LayeredMock {
+    type Error = core::convert::Infallible;
+
+    fn orchard_fvk(&self) -> Result<FullViewingKey, Self::Error> {
+        Ok(self.fvk.clone())
+    }
+
+    fn orchard_anchor(&self) -> Result<Anchor, Self::Error> {
+        Ok(self.anchor)
+    }
+
+    fn resolve_wallet_note(
+        &self,
+        index: usize,
+        _anchor: Anchor,
+    ) -> Result<(Note, MerklePath), Self::Error> {
+        Ok(self.witnesses[index].clone())
+    }
+
+    fn resolve_funding_notes(
+        &self,
+        values: &[u64],
+        _anchor: Anchor,
+    ) -> Result<Vec<(Note, MerklePath)>, Self::Error> {
+        // By-value greedy over the minted notes (index >= n_wallet), with a persistent used-set so
+        // successive resolutions (a layer's feeders, then a later layer's, then the funding notes)
+        // never reuse a note. Notes of equal value are interchangeable, so first-unused is correct.
+        let mut used = self.used.borrow_mut();
+        let mut out = Vec::with_capacity(values.len());
+        for &v in values {
+            let idx = (self.n_wallet..self.witnesses.len())
+                .find(|&i| !used[i] && self.witnesses[i].0.value().inner() == v)
+                .expect("a minted note of the requested value exists");
+            used[idx] = true;
+            out.push(self.witnesses[idx].clone());
+        }
+        Ok(out)
+    }
+
+    fn sign(&self, pczt: pczt::Pczt) -> Result<pczt::Pczt, Self::Error> {
+        Ok(sign_pczt(pczt, &self.ask).expect("signs the migration PCZT"))
+    }
+}
